+++ services/payments/tests/unit/test_rpc_payments.py
@@ -4,11 +4,10 @@
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
 
-from collections.abc import AsyncIterator, Awaitable, Callable
+from collections.abc import Awaitable, Callable
 
 import orjson
 import pytest
-from asgi_lifespan import LifespanManager
 from fastapi import FastAPI
 from models_library.api_schemas_webserver.wallets import WalletPaymentCreated
 from pytest_simcore.helpers.typing_env import EnvVarsDict
@@ -36,17 +35,6 @@ def app_environment(
             **rabbit_env_vars_dict,
         },
     )
-
-
-@pytest.fixture
-async def initialized_app(app: FastAPI) -> AsyncIterator[FastAPI]:
-    """Inits app on a light environment"""
-    async with LifespanManager(
-        app,
-        startup_timeout=10,
-        shutdown_timeout=10,
-    ):
-        yield app
 
 
 async def test_webserver_one_time_payment_workflow(
